# NodeJS/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "description": "Simple CRUD Task API built with Express.js",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Home",
        "description": "Returns API information.",
        "responses": {
          "200": {
            "description": "API information"
          }
        }
      }
    },

    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns server status.",
        "responses": {
          "200": {
            "description": "Server is healthy"
          }
        }
      }
    },

    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "description": "Returns all tasks.",
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        }
      },

      "post": {
        "summary": "Create a task",
        "description": "Creates a new task.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTask"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          },

          "400": {
            "description": "Invalid request"
          }
        }
      }
    },

    "/tasks/{id}": {

      "get": {

        "summary": "Get task by ID",

        "description": "Returns one task.",

        "parameters": [

          {

            "name": "id",

            "in": "path",

            "required": true,

            "schema": {

              "type": "integer"

            }

          }

        ],

        "responses": {

          "200": {

            "description": "Task found"

          },

          "404": {

            "description": "Task not found"

          }

        }

      },

      "put": {

        "summary": "Update task",

        "description": "Updates a task title or done status.",

        "parameters": [

          {

            "name": "id",

            "in": "path",

            "required": true,

            "schema": {

              "type": "integer"

            }

          }

        ],

        "requestBody": {

          "required": true,

          "content": {

            "application/json": {

              "schema": {

                "$ref": "#/components/schemas/UpdateTask"

              }

            }

          }

        },

        "responses": {

          "200": {

            "description": "Task updated"

          },

          "400": {

            "description": "Invalid request"

          },

          "404": {

            "description": "Task not found"

          }

        }

      },

      "delete": {

        "summary": "Delete task",

        "description": "Deletes a task.",

        "parameters": [

          {

            "name": "id",

            "in": "path",

            "required": true,

            "schema": {

              "type": "integer"

            }

          }

        ],

        "responses": {

          "204": {

            "description": "Task deleted"

          },

          "404": {

            "description": "Task not found"

          }

        }

      }

    }

  },

  "components": {

    "schemas": {

      "CreateTask": {

        "type": "object",

        "required": [

          "title"

        ],

        "properties": {

          "title": {

            "type": "string",

            "example": "Buy milk"

          }

        }

      },

      "UpdateTask": {

        "type": "object",

        "properties": {

          "title": {

            "type": "string",

            "example": "Finish Backend Assignment"

          },

          "done": {

            "type": "boolean",

            "example": true

          }

        }

      }

    }

  }

}